Implement thread pool for the Hexagon simulator.
