+++ test/correctness/dilate3x3.cpp
@@ -38,7 +38,7 @@ int main(int argc, char **argv) {
     if (target.has_gpu_feature()) {
         dilate3x3.gpu_tile(x, y, 16, 16);
     } else if (target.features_any_of({Target::HVX_64, Target::HVX_128})) {
-        dilate3x3.hexagon().vectorize(x, 64);
+        dilate3x3.hexagon().vectorize(x, 64).parallel(y);
     } else {
         dilate3x3.vectorize(x, target.natural_vector_size<uint8_t>());
     }
