+++ test/correctness/median3x3.cpp
@@ -54,7 +54,7 @@ int main(int arch, char **argv) {
     if (target.has_gpu_feature()) {
         median3x3.gpu_tile(x, y, 16, 16);
     } else if (target.features_any_of({Target::HVX_64, Target::HVX_128})) {
-        median3x3.hexagon().vectorize(x, 64);
+        median3x3.hexagon().vectorize(x, 64).parallel(y);
     } else {
         median3x3.vectorize(x, target.natural_vector_size<uint8_t>());
     }
